refactor(core): 遷移共用元件並優化全站圖片資源
本次提交主要包含兩大重構任務：共用元件的結構優化，以及全站圖片資源的載入效能提升。

- **決策**：建立 `components/shared/` 目錄，用於集中管理跨功能模組（如 `admin`, `company`, `users`）共用的基礎元件。
- **理由**：解決先前根目錄元件混雜的問題，提升專案結構的語意清晰度與長期可維護性。
- **實作**：
    - 遷移所有 Icon 元件至新目錄。
    - 遵循 Nuxt 3 的自動引入規則，系統性地更新了全站所有相關的元件標籤 (例如：`<BetaIcon>` -> `<SharedBetaIcon>`)。

- **決策**：將所有圖片資源從 `assets/` (建置期處理) 統一遷移至 `public/` (執行期處理)，並全面將原生 `<img>` 標籤升級為 `<NuxtImg>` 元件。
- **理由**：最大化利用 `@nuxt/image` 模組的執行期優化能力（如：自動格式轉換、尺寸裁切），以提升網站載入效能與使用者體驗。
- **實作**：
    - 遷移了 `assets/img/` 下的所有圖片資源。
    - 全面重構了 Vue 元件中的圖片引用方式，並修正了 CSS 中的 `url()` 路徑。

- 修復了在重構過程中，因自動化腳本錯誤修改 `<NuxtImg>` 元件而導致的渲染錯誤。
- 透過清除 `.nuxt` 快取並重啟開發伺服器，解決了因元件路徑變更所引發的 `DevalueError: Cannot stringify a function` 伺服器端渲染警告。
