+++ components/shared/BetaIcon.vue
