+++ components/shared/CloseIcon.vue
