+++ components/shared/LocationPinIcon.vue
