+++ components/shared/StarIcon.vue
